add profiler
